Move page size calculator to layouts package
