+++ src/Viewer.tsx
@@ -13,11 +13,11 @@ import defaultLayout from './layouts/defaultLayout';
 import defaultToolbar from './layouts/defaultToolbar';
 import Inner from './layouts/Inner';
 import { Layout } from './layouts/Layout';
+import PageSizeCalculator, { PageSize } from './layouts/PageSizeCalculator';
 import { RenderToolbar } from './layouts/ToolbarSlot';
 import DocumentLoader from './loader/DocumentLoader';
 import LocalizationMap from './localization/LocalizationMap';
 import LocalizationProvider from './localization/LocalizationProvider';
-import PageSizeCalculator, { PageSize } from './PageSizeCalculator';
 import PdfJs from './PdfJs';
 import SelectionMode from './SelectionMode';
 import ThemeProvider from './theme/ThemeProvider';
